project01/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
crispy-bootstrap4==2022.1
Django==4.2.2
django-bootstrap-v5==1.0.11
django-crispy-forms==2.0
djangorestframework==3.14.0
psycopg2-binary==2.9.6
pytz==2023.3
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.6.3
